{% include 'clonotypr/navbar.html' %}

File upload by directory for AddClonotype

AddByDir uploads reads in files needed by the AddClonotype function.

AddByDir(directory)

Arguments

directory
Directory containing both "clonotypes.csv" and "filtered_contig_annotations.csv".

Value

list; returns two dataframes generated from reading in both CSV files

Details

The function is provides one of the optional upload strategies for AddClonotype. It scans the directory for "clonotypes.csv" and "filtered_contig_annotations.csv", then reas in both and returns both.

See also

Examples

AddByDir(directory = "data")

#> Error in AddByDir(directory = "data"): could not find function "AddByDir"

AddByDir(directory = "data/")

#> Error in AddByDir(directory = "data/"): could not find function "AddByDir"