{% include 'clonotypr/navbar.html' %}

Generate ClonoSeurat Objects

AddClonotype attaches clonotype data to a Seurat object as metadata

AddClonotype(object, directory, clonotypeFile, filteredContigFile)

Arguments

object
A Seurat object created with CreateSeuratObject

directory
Directory containing clonotypes.csv and filtered_contig_annotations.csv

clonotypeFile
CSV file containing CDR3 and clonotype IDs from 10X outs. Default ilename is "clonotypes.csv"

filteredContigFile
CSV file containing cell barcodes and clonotype IDs from 10X outs. Default filename is "filtered_contig_annotations.csv"

Value

object; Seurat object with clonotype data appended to Seurat meta.data

Details

The function appends CSV clonotype data generated from CellRanger outs to Seurat objects. The two required files have default names of "clonotypes.csv" and "filtered_contig_annotations.csv". Two options for the function call are possible: 1. SeuratObject + the directory containing the files or 2. SeuratObject + "clonotypes.csv" + "filtered_contig_annotations.csv". See "Examples" for possible function calls.

See also

Examples

clonotypes <- system.file("extdata", "vdj_v1_hs_pbmc3_t_clonotypes.csv", package = "clonotypr")
fca <- system.file("extdata", "vdj_v1_hs_pbmc3_t_filtered_contig_annotations.csv", package = "clonotypr")
pbmc_vdj <- AddClonotype(object = pbmc_null, clonotypeFile = clonotypes, filteredContigFile = fca)