{% include 'clonotypr/navbar.html' %}

Build CDR3 Composition Matrices

AnalyzeCDR3aa divides sequences by length and parses the seqeunce strings. Then can generate either a position frequency matrix (PFM), position probability matrix (PPM), or probability weight matrix (PWM). See details for more.

AnalyzeCDR3aa(data, chain, pseudocount = 1, output = "PPM")

Arguments

data
Seurat object or dataframe with cdr3 chains separated and parsed

chain
Lymphocyte-receptor chain. Specify receptor chain as it appears in MeltCDR3Lengths (i.e. "TRA"). For B-Cell isotypes, such as IgA, specify chain="IGH" to isolate the isotypes

pseudocount
Offset to avoid matrix entries == 0. Default = 1. Value of 1 sets offset = to 0.05; offset = (1 / # of AAs). Set `pseudocount = 0`` to generate untouched matrices.

output
Calculation to be performed. Must be either "PFM", "PPM", or "PWM". Default set to PPM

Value

AnalyzeCDR3aa returns a list of dataframes for the given output. List contains the specified output for each sequence length of the specified chain.

Details

PFMs can be calculated here to understand the frequency of each AA, at each position, in all CDR3 sequences of a specific sequence length.

To calculate the PPM, the frequencies are normalized by the total sequence length for a given PFM. To avoid values of 0 (due to either small sample or less likely AA at given position -- i.e anchor residues), we can use a pseudocount to keep all values in the PPM greater than 0.

The PPM is used to calculate the PWM, by normalizing the PPM to a background frequency. AnalyzeCDR3aa defaults to 1 for this value.

** A digestable eplanation of PFM, PPM, and PWM can be found on here: https://en.wikipedia.org/wiki/Position_weight_matrix

See also

Examples

pos_weight_matrices <- AnalyzeCDR3aa(data=pbmc_vdj, chain="TRA", pseudocount=1, output="PWM")

#> Error in x[[jj]] <- v: attempt to select less than one element in integerOneIndex

pos_weight_matrices[[3]]

#> Error in eval(expr, envir, enclos): object 'pos_weight_matrices' not found