{% include 'clonotypr/navbar.html' %}

Build Position Probability Matrices

BuildPPM calculates position probability matrices for CDR3 chains separated by CDR3 sequence length using a given pseudocount

BuildPPM(data, chain, pseudocount = 1, lengthIndex)

Arguments

data
Seurat object with clonotype metadata or dataframe containing columns labeled "alpha"/"beta" and/or "alpha.length"/"beta.length".

chain
Lymphocyte-receptor chain. Specify receptor chain as it appears in MeltCDR3Lengths (i.e. "TRA"). For B-Cell isotypes, such as IgA, specify chain="IGH" to isolate the isotypes

pseudocount
Offset to avoid matrix entries == 0. Default == 1. Value of 1 sets offset equal to 0.05; offset = (1 / # of AAs). Set `pseudocount = 0`` to generate untouched matrices.

lengthIndex
AA length index in list of c(min(aa_length) to max(aa_length))

Value

BuildPPM returns a position probability matrix as a dataframe

Details

BuildPPM has been integrated into link{AnalyzeCDR3aa} and will be depracated in later updates. Please use AnalyzeCDR3aa(data=data, chain="alpha", pseudocount=1, output="PPM")

Refer to https://en.wikipedia.org/wiki/Position_weight_matrix for general information about PFMs, PPMs, and PWMs

See also

Examples

matrix <- BuildPPM(pbmc_vdj, "alpha", pseudocount=1, 5)

#> Error in BuildPFM(data = data, chain = chain, lengthIndex = lengthIndex): Chain option is not offered. Check the documentation

matrix

#> function (data = NA, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL) 
#> {
#>     if (is.object(data) || !is.atomic(data)) 
#>         data <- as.vector(data)
#>     .Internal(matrix(data, nrow, ncol, byrow, dimnames, missing(nrow), 
#>         missing(ncol)))
#> }
#> <bytecode: 0x7f8dd6317ec0>
#> <environment: namespace:base>