{% include 'clonotypr/navbar.html' %}

Format dataframes in AddClonotype

FormatClonoDataframe processes the dataframes in the AddClonotype function.

FormatClonoDataframe(clonoTable, clonoCSV)

Arguments

clonoTable
dataframe generated from the filtered_contig_annotations.csv file

clonoCSV
dataframe generated from the clonotypes.csv file

Value

list; returns two dataframes generated from reading in both CSV files

Details

Dataframes are returned from either AddByFile or AddByDir and processed with this function. The two dataframes are merged by this and formatted before AddClonotype appends them to the SeuratObject metadata

See also

Examples

FormatClonoDataframe(clonoTable = dataframe1, clonoCSV = dataframe2)

#> Error in FormatClonoDataframe(clonoTable = dataframe1, clonoCSV = dataframe2): could not find function "FormatClonoDataframe"