{% include 'clonotypr/navbar.html' %}

Find Differentially Expressed Genes

A streamlined way to return genes that are differentially expressed between Seurat subsets.

GetDEGenes(object, subset1 = NULL, subset2 = NULL)

Arguments

object
A Seurat object created with CreateSeuratObject

subset1
Subset created from a Seurat object; class(subset1) == "Seurat"

subset2
(optional) Subset created from a Seurat object; class(subset2) == "Seurat"

Value

GetDEGenes returns the same output as FindMarkers

Details

Two options are available for calling GetDEGenes: 1.) Passing in two unique subsets and finding DE genes between the two 2.) Passing in only 1 subset and retrieving genes that are differentially expressed between the subset and the total cell population

See also

Examples

cd4PosCells <- subset(pbmc_vdj, subset=CD4 > 0)
cd8PosCells <- subset(pbmc_vdj, subset=CD8A > 0)
de_genes <- GetDEGenes(object = pbmc_vdj, subset1=cd3PosCells, subset2=cd8PosCells)

#> Error in GetDEGenes(object = pbmc_vdj, subset1 = cd3PosCells, subset2 = cd8PosCells): could not find function "GetDEGenes"

de_genes <- GetDEGenesfunction(object = pbmc_vdj, subset1=cd3PosCells, subset2=NULL)

#> Error in GetDEGenesfunction(object = pbmc_vdj, subset1 = cd3PosCells,     subset2 = NULL): could not find function "GetDEGenesfunction"