{% include 'clonotypr/navbar.html' %}

Add single chain clonotypes to freq table

JoinSingleChains generates a frequency table similiar to GetFrequencyDF, but adds the frequency of single chain clonotypes to their corresponding productive clonotypes. Handles both CDR3 alpha and CDR3 beta single chains.

JoinSingleChains(data)

Arguments

data
Dataframe or Seurat data with metadata containing clonotype data usingLoadClonotypes or AddClonotypes, respectively.

Value

Returns a frequency dataframe containing only productive sequences and their frequencies

See also

Examples

prod_freqs <- JoinSingleChains(data = pbmc_vdj)

#> Error in AnalyzeCDR3Lengths(barcodeFreqs, productive = FALSE, beta = FALSE,     alpha = TRUE): unused arguments (productive = FALSE, beta = FALSE, alpha = TRUE)