{% include 'clonotypr/navbar.html' %}

Format Clonotype Dataframe From CellRanger Files

LoadClonotypes reads in clonotype data from the 'clonotypes summary CSV' and 'filtered contig annotations' files and formats them into a dataframe.

LoadClonotypes(directory, clonotypeFile, filteredContigFile, cdr3 = TRUE)

Arguments

directory
(option #1) Directory containing clonotypes.csv and filtered_contig_annotations.csv

clonotypeFile
(option #2)CSV file containing CDR3 and clonotype IDs from 10X outs. Default ilename is "clonotypes.csv"

filteredContigFile
(option #2)CSV file containing cell barcodes and clonotype IDs from 10X outs. Default filename is "filtered_contig_annotations.csv"

cdr3
Logical. Include analysis of cdr3s_aa & cdr3s_nt. Essentially running AnalyzeCDR3Lengths internally

Value

object; Seurat object with clonotype data appended to Seurat meta.data

Details

The function builds and formats a dataframe from raw clonotype-related CSVs generated from CellRanger. The two required files have names similarto "clonotypes.csv" and "filtered_contig_annotations.csv" (see example). Two options are available for the function arguments: 1. The directory containing the files or 2. "ClonotypesSummaryFile.csv" + "filtered_contig_annotations.csv".

See also

Examples

clonotypes <- system.file("extdata", "vdj_v1_hs_pbmc3_t_clonotypes.csv", package = "clonotypr")
fca <- system.file("extdata", "vdj_v1_hs_pbmc3_t_filtered_contig_annotations.csv", package = "clonotypr")
clonotypes <- LoadClonotypes(clonotypeFile = clonotypes, filteredContigFile = fca)