{% include 'clonotypr/navbar.html' %}

Plot CDR3 Chain Composition

PlotCDR3aa plots the CDR3 amino acid composition of alpha, beta, or both chains in a dataframe containing clonotypes The default strategy is to use the output from AnalyzeCDR3Lengths with PlotCDR3aa.

PlotCDR3aa(
  data,
  chain,
  length = "all",
  palette = "bold",
  cols = 3,
  method = "prob",
  theme = "pubr",
  bg.color = "light"
)

Arguments

data
Can be a dataframe or Seurat object. Must contain "cdr3s_aa" and/or "cdr3s_nt" columns

chain
Lymphocyte-receptor chain. Specify receptor chain as it appears in MeltCDR3Lengths (i.e. "TRA"). For B-Cell isotypes, such as IgA, specify chain="IGH" to isolate the isotypes

length
Default "all" will generate plots for CDR3alpha and CDR3beta chains of all lengths. If length is set to a numeric sequence length, a single plot will be generated -- alpha or beta must also be set to TRUE if using a numeric.

palette
Any ggsci, color brewer, or palettes available from PullPalettes

cols
Number of columns in each plot layout

method
Potting method. Either 'bits' or 'prob' (probability)

theme
Either ggpubr::theme_pubr("pubr"), or ggpubr::pubclean ("clean"), or ggpubr::theme_transparent ("transparent")

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

Value

Using the default settings length="all" will returns list(CDR3_alphas, CDR3_betas). If length is set to a numeric length=15 and alpha = TRUE or beta = TRUE, a single plot will be returned. All outputs can be customized with standard ggplot params

See also

Examples

PlotCDR3aa(data=pbmc_vdj, chain="TRA", length=12, color='auto', cols=1, method='prob')

#> Error in PlotCDR3aa(data = pbmc_vdj, chain = "TRA", length = 12, color = "auto",     cols = 1, method = "prob"): unused argument (color = "auto")