{% include 'clonotypr/navbar.html' %}

Plot Distribution Of Clonotype Frequencies

PlotClonotypeDist plots the distribution of clonotype frequencies. Frequency bins are predefined to focus on rare populations.

PlotClonotypeDist(
  data,
  palette = "bold10",
  title = TRUE,
  x.text.angle = 45,
  barOutline = TRUE,
  legend = TRUE,
  gridlines = TRUE,
  alpha = 0.5,
  label = TRUE,
  theme = "pubr",
  bg.color = "light"
)

Arguments

data
Can be a dataframe with clonotype IDs as "clonotype_id" or frequencies as "Freq". Can also be a Seurat object containing with either of the same column names and data

palette
Palettes available in ggplot2, ggsci, or with PullPalettes. Default = "bold10"

title
Logical. Whether to include plot title or not

x.text.angle
Numerical. Angle of the text labels on the x-axis. Default = 45

barOutline
Logical. Whether to include a black outline around the bars in the plot

legend
Logical. Whether to include the legend or not

gridlines
Logical. Include y-axis

alpha
Numerical between 0-1. Alpha level applied to bar-fill.

label
Logical. Add numerical label to the top of the bars

theme
Either ggpubr::theme_pubr("pubr"), or ggpubr::pubclean ("clean"), or ggpubr::theme_transparent ("transparent")

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

Value

PlotClonotypeDist returns a single ggplot barplot object.

See also

Examples

plot <- PlotClonotypeDist(data = pbmc_vdj)