{% include 'clonotypr/navbar.html' %}

Plot Histogram of All Genes in Chain Region

PlotIgIsotypes Plots the isotype counts using data from CountIgIsotypes

PlotIgIsotypes(
  isotype_counts,
  plot.type = "bar",
  bar.type = "stack",
  palette = "bold",
  barOutline = FALSE,
  gridlines = TRUE,
  alpha = 0.5,
  theme = "pubr",
  bg.color = "light"
)

Arguments

isotype_counts
Data from CountIgIsotypes

plot.type
Plot as either a barplot or balloon plot

bar.type
Type of barplot. Options == c("stack", "dodge", "facet")

palette
Any ggsci, color brewer, or palettes available from PullPalettes

barOutline
Logical. Whether to include a black outline around the bars in the plot

gridlines
Logical. Include y-axis

alpha
Numerical between 0-1. Alpha level applied to bar-fill.

theme
Either ggpubr::theme_pubr("pubr"), or ggpubr::pubclean ("clean"), or ggpubr::theme_transparent ("transparent")

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

x.text.angle
Numerical. Angle of the text labels on the x-axis. Default = 45

Value

PlotIgIsotypes returns a single ggplot object

See also

Examples

isotype_counts <- CountIgIsotypes(pbmc_vdj[[]])

#> Error: object 'cell_type' not found

plot <- PlotIgIsotypes(isotype_counts, plot.type="bar", palette="bold")

#> Error in ggpubr::ggbarplot(data = isotype_counts, x = "gene", y = "count",     order = c("IGHG", "IGHA", "IGHM", "IGHE", "IGHD"), xlab = "Isotype",     ylab = "Count", label = TRUE, lab.col = lab.col, color = "chain",     fill = "chain", palette = palette, alpha = 0.7, position = position,     facet.by = facet, size.range = c(1, 20), ggtheme = theme): object 'isotype_counts' not found

plot

#> function (x, y, ...) 
#> UseMethod("plot")
#> <bytecode: 0x7f8dd6979ab0>
#> <environment: namespace:graphics>