{% include 'clonotypr/navbar.html' %}

Plot Isotype Counts For Ig Data

PlotIsotypeCounts Merges isotypes together using data from AnalyzeGenes and plots the data either as a donut plot or bar plot.

PlotIsotypeCounts(
  gene_data,
  plot.type = "bar",
  palette = "bold",
  barOutline = FALSE,
  interactive = FALSE,
  gridlines = TRUE,
  alpha = 0.5,
  label = TRUE,
  theme = "pubr",
  bg.color = "light"
)

Arguments

gene_data
Data from running AnalyzeGenes

plot.type
Plot as either a barplot or donut plot

palette
Any ggsci, color brewer, or palettes available from PullPalettes

barOutline
Logical. Whether to include a black outline around the bars in the plot

interactive
Logical. Plot interactive plotly plots or flat plots

gridlines
Logical. Include y-axis

alpha
Numerical between 0-1. Alpha level applied to bar-fill.

label
Logical. Add numerical label to the top of the bars

theme
Either ggpubr::theme_pubr("pubr"), or ggpubr::pubclean ("clean"), or ggpubr::theme_transparent ("transparent")

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

Value

PlotIsotypeCounts returns plot of isotype counts for Ig data

Details

Both plot options offered as flat or interactive plots using the interactive param. Available palettes are pulled from PullPalettes

See also

Examples

plot <- PlotIsotypeCount(genes, plot.type = "donut", palette="new_england", interactive=TRUE)

#> Error in PlotIsotypeCount(genes, plot.type = "donut", palette = "new_england",     interactive = TRUE): could not find function "PlotIsotypeCount"

plot

#> function (x, y, ...) 
#> UseMethod("plot")
#> <bytecode: 0x7f8dd6979ab0>
#> <environment: namespace:graphics>