{% include 'clonotypr/navbar.html' %}

Plot CDR3 AA Length Distributions

PlotLengths plots the distribution of amino acid lengths for both the alpha and beta CDR3 sequences.

PlotLengths(data)

Arguments

data
Can be a dataframe with CDR3 length data or Seurat object containing at least two columns: "cdr3s_aa"and "clonotype_id". If Seurat object, columns should be in the meta.data slot.

Value

PlotLengths returns two plots in a list. list[1] is a plot of the CDR3 alpha length distribution. list[2] is a plot of the CDR3 beta distribution. Both plots are ggplot objects and can be customized using standard the ggplot2 library

See also

Examples

plots <- PlotLengths(data = pbmc_vdj)