{% include 'clonotypr/navbar.html' %}

Plot CDR3 Length Distributions

PlotLengthsByChain plots the distribution of amino acid lengths for CDR3 sequences

PlotLengthsByChain(
  data,
  cell.type = "T",
  plot.type = "stack",
  cdr3 = "aa",
  palette = "bold",
  barOutline = FALSE,
  legend = TRUE,
  gridlines = TRUE,
  alpha = 0.5,
  label = TRUE,
  theme = "pubr",
  bg.color = "light"
)

Arguments

data
cdr3_melt data created from MeltCDR3Lengths

cell.type
Lymphocyte To filter on. Options == c("T", "B", or "all")

plot.type
Type of barplot. Options == c("stack", "dodge", "facet")

cdr3
CDR3 type. Options == c("aa", "nt")

palette
Palettes available from PullPalettes

barOutline
Logical. Whether to include a black outline around the bars in the plot

legend
Logical. Whether to include the legend or not

gridlines
Logical. Include y-axis

alpha
Numerical between 0-1. Alpha level applied to bar-fill.

label
Logical. Add numerical label to the top of the bars

theme
Either ggpubr::theme_pubr("pubr"), or ggpubr::pubclean ("clean"), or ggpubr::theme_transparent ("transparent")

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

Value

PlotLengthsByChain

See also

Examples

melt_data <- MeltCDR3Lengths(data = pbmc_vdj)

#> Error in `[.Seurat`(data, c(stringb::text_detect(colnames(data), "length") &     c(stringb::text_detect(colnames(data), "aa")))): Incorrect number of logical values provided to subset features

plot <- PlotLengths(data = melt_data)

#> Error in PlotLengths(data = melt_data): could not find function "PlotLengths"