{% include 'clonotypr/navbar.html' %}

Plot Clonotype Productivity Counts

PlotProductivity plots productivity counts for clonotype sample as donut plot

PlotProductivity(
  data,
  interactive = FALSE,
  palette = "bold",
  alpha = 0.7,
  theme = "pubr",
  bg.color = "light"
)

Arguments

data
Seurat object with clonotype data or dataframe. Use after AddClonotypes or LoadClonotypes with cdr3=TRUE

interactive
Logical. Whether to plot as ggplot or plotly

palette
Palette pulled from PullPalettes or any ggsci or ColorBrewer palette

alpha
Numerical between 0-1. Alpha level applied to section-fill.

theme
Either ggpubr::theme_pubr("pubr"), or ggpubr::pubclean ("clean"), or ggpubr::theme_transparent ("transparent")

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

Value

PlotProductivity returns a donut chart

Examples