{% include 'clonotypr/navbar.html' %}

Plot Network Graph To Show VDJ Gene Relationships

PlotVDJByNetwork Plots the weighted relationship between VDJ genes across the cell sample. Plot is either a Sankey diagram or Chord diagram

PlotVDJByNetwork(
  edges,
  nodes,
  connections,
  type = "sankey",
  segment,
  pair = "VJ",
  palette = "bold",
  bg.color = "light"
)

Arguments

edges
Edge list, created from BuildVDJNetwork

nodes
Node list, created from BuildVDJNetwork

connections
Connections list, created from BuildVDJNetwork

type
Type of plot. Either "sankey" or "chord"

segment
VDJ gene for chain of immune cell receptor (i.e. "TRAV9")

pair
Pair of VDJ relationships. Ex.) 'VD' or 'VJ', or 'DJ'

palette
Any ggsci, color brewer, or palettes available from PullPalettes

bg.color
Helpful param for applying a global light or dark theme in a scheme, such as global theme for Shiny apps. "light" or "dark".

Value

PlotVDJByNetwork returns a sankey diagram or chord diagram depending on type param

See also

Examples

gene_network <- BuildVDJNetwork(data=pbmc_vdj[[]], chain="TRB")

#> Error in BuildVDJNetwork(data = pbmc_vdj[[]], chain = "TRB"): unused argument (chain = "TRB")

edges <- gene_network[[1]]

#> Error in eval(expr, envir, enclos): object 'gene_network' not found

nodes <- gene_network[[2]]

#> Error in eval(expr, envir, enclos): object 'gene_network' not found

connections <- gene_network[[3]]

#> Error in eval(expr, envir, enclos): object 'gene_network' not found

sankey <- PlotVDJByNetwork(edges, nodes, connections,type="sankey")

#> Error in dplyr::mutate(nodes, id = id - 1): object 'nodes' not found

sankey

#> Error in eval(expr, envir, enclos): object 'sankey' not found