{% include 'clonotypr/navbar.html' %}

Find Clonotypes Matching Filters

WhichClonotypes parses through all available clonotypes and returns a subset which match the gene filters applied. Builds off of Seurat::WhichCells

WhichClonotypes(object, filters)

Arguments

object
A Seurat object with clonotype data attached via AddClonotypes

filters
Conditional gene filter statements. Can be passed as a single character statement or as multiple. When passing in multiple gene conditionals, combine as one character string: i.e "CD3E > 0 & CD8A > 0"

Value

WhichClonotypes returns clonotypes that meet the gene filter criteria

See also

Examples

clonotypes <- WhichClonotypes(object = pbmc_vdj, filters = "CD3E > 0")
clonotypes <- WhichClonotypes(object = pbmc_vdj, filters = "CD3E > 0 & CD8A > 0")